third_party/WebKit/LayoutTests/fast/canvas-api/canvas-overloads-strokeText.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
var ctx = document.createElement('canvas').getContext('2d');

function ExpectedMessage(num) {
    return "\"TypeError: Failed to execute 'strokeText' on 'CanvasRenderingContext2D': 3 arguments required, but only " + num + " present.\"";
}

test(function(t) {
    try {
      ctx.strokeText();
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
      assert_equals(ex.message, "Failed to execute 'strokeText' on 'CanvasRenderingContext2D': 3 arguments required, but only 0 present.");
    }
    try {
      ctx.strokeText('moo');
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
      assert_equals(ex.message, "Failed to execute 'strokeText' on 'CanvasRenderingContext2D': 3 arguments required, but only 1 present.");
    }
    try {
      ctx.strokeText('moo', 0);
    } catch (ex) {
      assert_equals(ex.name, "TypeError");
      assert_equals(ex.message, "Failed to execute 'strokeText' on 'CanvasRenderingContext2D': 3 arguments required, but only 2 present.");
    }
    assert_equals(ctx.strokeText('moo',0,0), undefined);
    assert_equals(ctx.strokeText('moo',0,0,0), undefined);
    assert_equals(ctx.strokeText('moo',0,0,0,0), undefined);
}, "Test the behavior of CanvasRenderingContext2D.strokeText() when called with different numbers of arguments.");
</script>
</body>
